UEFI: Fastboot: Reduce HiKey max download size to avoid failures after chainloading from grub
When fastboot.efi is chainloaded from grub, we end up with not
enough contiguous memory to allocate 500M chunks. Sadly we also
can't allocate 256M chunks either (which is a bit silly).

None the less, this patch provides a workaround, limiting the
max size to 128M chunks. This does not cause flashing to take
more time, but does spit out more output when flashing the
multiple chunks.

Signed-off-by: John Stultz <john.stultz@linaro.org>