DwSdDxe: Use a timeout counter to avoid uSD hang waiting for idle
This is a tweaked version of a patch earlier applied by Haojian.

In some cases, we're seeing bootup hang while polling SD_STATUS register,
and it seems to be stuck with the DATA_BUSY signal.

This patch tries to resolve this quirk by using a max loop counter
which will break out of the wait for idle loops if they run for
too long.

This can introduce small delays on boot when a uSD card is present
but is better then a hang, and avoids potential issues where we proceed
before the card is ready if the loop was just removed.

This patch also addresses both wait-for-idle loops.

Change-Id: I4be39e11eb42b611ab4ae50c4bc42324b00de9ee
Signed-off-by: John Stultz <john.stultz@linaro.org>