HiKeyFastboot.c: add support for sparse image with CHUNK_TYPE_FILL
since the sparse image generated with img2simg command will have
CHUNK_TYPE_FILL type trunk, and it's better to have support for that,
so that we can try other file system type with img2simg command.

And we will ignore the FILL type chunk if it is filled with 0.

If the it's not filled with 0, it will need more time to finish the flash

Change-Id: I33ec90dbb0255b59cf5afcad67061fa7a89be106
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>